backend/requirements.txt
Django>=4.2.0,<5.0
djangorestframework
djangorestframework-simplejwt[crypto]>=5.3.0
django-filter>=23.5
django-cleanup
psycopg[binary]>=3.1.0
python-decouple
djangorestframework-simplejwt
django-cors-headers
qrcode[pil]
qrcode[pil]>=7.4.2
Pillow>=10.0.0
pytest>=7.4
pytest-django>=4.5
factory-boy>=3.3
celery[redis]>=5.3.0
django-celery-beat>=2.5.0
redis>=4.5.0
requests
python-decouple
channels==4.1.0
channels_redis==4.2.0
